tests/submission/Opera/script_scheduling/112.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: removing async attribute at runtime, script also has defer attribute</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="test"></div>
	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

        <script id="async-script" async defer src="scripts/delay.php?url=include-1.js&time=1"></script> 

	<script>
          var t = async_test();

          addEventListener("DOMContentLoaded", t.step_func(function () {
            assert_array_equals(eventOrder, []);
            t.done();
          }), false);

          addEventListener("load", t.step_func(function () {
            assert_array_equals(eventOrder, ["external script #1"]);
            t.done();
          }), false);
          
        </script>
</body></html>
